fix: avoid c++ narrowing and time conversion warnings

+++ src/slmp_minimal.cpp
@@ -17,7 +17,9 @@
  */
 static uint32_t millis() {
     auto now = std::chrono::steady_clock::now();
-    return std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count();
+    return static_cast<uint32_t>(
+        std::chrono::duration_cast<std::chrono::milliseconds>(now.time_since_epoch()).count()
+    );
 }
 #endif
